# Optimize AbstractTaggedMetricRegistry#getOrAdd for read path