--- a/tritium-jmh/build.gradle
+++ b/tritium-jmh/build.gradle
@@ -4,7 +4,7 @@ jmh {
     resultFormat = 'JSON'
     // Use profilers to collect additional data. Supported profilers:
     // [cl, comp, gc, stack, perf, perfnorm, perfasm, xperf, xperfasm, hs_cl, hs_comp, hs_gc, hs_rt, hs_thr]
-    profilers = ['gc']
+    profilers = ['gc', 'jfr']
 }
 
 tasks.jmhCompileGeneratedClasses {