--- a/tritium-registry/src/main/java/com/palantir/tritium/metrics/registry/AbstractTaggedMetricRegistry.java
+++ b/tritium-registry/src/main/java/com/palantir/tritium/metrics/registry/AbstractTaggedMetricRegistry.java
@@ -214,7 +214,11 @@ public abstract class AbstractTaggedMetricRegistry implements TaggedMetricRegist
 
     protected final <T extends Metric> T getOrAdd(
             MetricName metricName, Class<T> metricClass, Supplier<T> metricSupplier) {
-        Metric metric = registry.computeIfAbsent(metricName, _name -> metricSupplier.get());
+        Metric metric = registry.get(metricName);
+        if (metric == null) {
+            // assume optimistic read fast path to avoid lamdba capture and fallback to slow write path
+            metric = registry.computeIfAbsent(metricName, _name -> metricSupplier.get());
+        }
         return checkNotNull(checkMetricType(metricName, metricClass, metric), "metric");
     }