## Before this PR
Some JFR profiles were showing excessive allocations of `com.palantir.tritium.metrics.registry.AbstractTaggedMetricRegistry$$Lambda$124+0x0000000800da2fa8.472991420` via `java.lang.invoke.Invokers$Holder.linkToTargetMethod(Object, Object)` from `AbstractTaggedMetricRegistry.getOrAdd(MetricName, Class, Supplier)`

![image](https://user-images.githubusercontent.com/54594/156796885-88a83107-92c5-4bc6-a533-2aac9fa33b53.png)


## After this PR
<!-- User-facing outcomes this PR delivers go below -->
==COMMIT_MSG==
Optimize AbstractTaggedMetricRegistry#getOrAdd for read path

Add MetricsLifecycleBenchmark.
JMH benchmarks produce JSON & JFR outputs
==COMMIT_MSG==

[JMH comparison](https://jmh.morethan.io/)


## Possible downsides?
<!-- Please describe any way users could be negatively affected by this PR. -->