@schlosna 2022-03-04
After running benchmarks across JDK 11, 15 and 17, I don't think this is actually worth modifying `getOrAdd`, so I'm going to close this out.

```
JDK 17.0.1, OpenJDK 64-Bit Server VM, 17.0.1+12-LTS

Before
Benchmark                            Mode  Cnt     Score      Error  Units
MetricsLifecycleBenchmark.counter    avgt    3   689.290 ± 8172.421  ns/op
MetricsLifecycleBenchmark.histogram  avgt    3   828.058 ± 3576.649  ns/op
MetricsLifecycleBenchmark.meter      avgt    3   813.318 ± 3006.805  ns/op
MetricsLifecycleBenchmark.timer      avgt    3  1375.768 ± 3100.411  ns/op

After
Benchmark                            Mode  Cnt     Score      Error  Units
MetricsLifecycleBenchmark.counter    avgt    3   734.545 ± 9357.108  ns/op
MetricsLifecycleBenchmark.histogram  avgt    3   858.349 ± 4133.059  ns/op
MetricsLifecycleBenchmark.meter      avgt    3   851.431 ± 2746.278  ns/op
MetricsLifecycleBenchmark.timer      avgt    3  1523.786 ± 9908.532  ns/op
```